JBTM-3934 move client API into an internal package
