+++ rts/lra/client/src/main/java/io/narayana/lra/client/internal/NarayanaLRAClient.java
@@ -3,7 +3,7 @@
    SPDX-License-Identifier: Apache-2.0
  */
 
-package io.narayana.lra.client;
+package io.narayana.lra.client.internal;
 
 import io.narayana.lra.Current;
 import io.narayana.lra.LRAConstants;
@@ -29,7 +29,6 @@ import jakarta.ws.rs.core.MultivaluedHashMap;
 import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.UriBuilder;
-import jakarta.ws.rs.core.UriInfo;
 import org.eclipse.microprofile.lra.annotation.AfterLRA;
 import org.eclipse.microprofile.lra.annotation.Compensate;
 import org.eclipse.microprofile.lra.annotation.Complete;
