+++ rts/lra/test/basic/src/test/java/io/narayana/lra/arquillian/TestBase.java
@@ -7,7 +7,7 @@ package io.narayana.lra.arquillian;
 
 import io.narayana.lra.LRAConstants;
 import io.narayana.lra.LRAData;
-import io.narayana.lra.client.NarayanaLRAClient;
+import io.narayana.lra.client.internal.NarayanaLRAClient;
 import jakarta.json.Json;
 import jakarta.json.JsonArray;
 import jakarta.json.JsonReader;
