+++ rts/lra/test/basic/src/test/java/io/narayana/lra/arquillian/client/LRAIT.java
@@ -8,7 +8,7 @@ package io.narayana.lra.arquillian.client;
 import io.narayana.lra.arquillian.Deployer;
 import io.narayana.lra.arquillian.TestBase;
 import io.narayana.lra.arquillian.resource.LRAParticipant;
-import io.narayana.lra.client.NarayanaLRAClient;
+import io.narayana.lra.client.internal.NarayanaLRAClient;
 import org.eclipse.microprofile.lra.annotation.LRAStatus;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.test.api.ArquillianResource;
