+++ rts/lra/test/basic/src/test/java/io/narayana/lra/arquillian/resource/LRAParticipant.java
@@ -5,7 +5,7 @@
 
 package io.narayana.lra.arquillian.resource;
 
-import io.narayana.lra.client.NarayanaLRAClient;
+import io.narayana.lra.client.internal.NarayanaLRAClient;
 import org.eclipse.microprofile.lra.annotation.Compensate;
 import org.eclipse.microprofile.lra.annotation.LRAStatus;
 import org.eclipse.microprofile.lra.annotation.ws.rs.LRA;
